Added a variant of the program using a structure.
The structure is used to hold the maximum possible value.
A double with size 32 bits and another byte for the carry.
In total 40 bits are used to hold the data.
The maximum possible value is on 33 bits.
Another idea would be to try to implement the result on 33 bits.
One way would be to use structures. (If it works)